Convert long ids to Integer

+++ src/main/java/com/soup/exambyte/models/Question.java
@@ -5,7 +5,7 @@ package com.soup.exambyte.models;
  */
 public abstract class Question {
 
-  private Long id;
+  private Integer id;
   private String questionTitle;
   private String questionDescription;
 
@@ -16,7 +16,7 @@ public abstract class Question {
    * @param questionDescription The description of the question.
    */
   public Question(String questionTitle, String questionDescription) {
-    this.id = (long) (Math.random() * 10);
+    this.id = (int) (Math.random() * 10);
     this.questionTitle = questionTitle;
     this.questionDescription = questionDescription;
   }
@@ -37,11 +37,11 @@ public abstract class Question {
     this.questionDescription = questionDescription;
   }
 
-  public Long getId() {
+  public Integer getId() {
     return id;
   }
 
-  public void setId(Long id) {
+  public void setId(Integer id) {
     this.id = id;
   }
 }
